+++ src/types/comment.types.ts
@@ -1,0 +1,7 @@
+export interface CommentType {
+  idx: number;
+  content: string;
+  nickname: string;
+  image: string;
+  isMine: boolean;
+}
